+++ Gamnet/Network/Router/RouterCaster.h
@@ -61,7 +61,7 @@ struct RouterCaster
 {
 private :
 	std::atomic<uint32_t> msg_seq;
-	std::shared_ptr<RouterCasterImpl> arrCasterImpl_[ROUTER_CAST_TYPE::MAX];
+	std::shared_ptr<RouterCasterImpl> arrCasterImpl_[(int)ROUTER_CAST_TYPE::MAX];
 
 public :
 	RouterCaster();
